Super-admin e autorização do delete e update

+++ app/Http/Controllers/ConteudoController.php
@@ -13,7 +13,7 @@ use App\Traits\FileSystemLogic;
 use Illuminate\Support\Facades\Mail;
 use App\Mail\MailUsuario;
 use App\User;
-use Illuminate\Support\Facades\Gate;
+use Gate;
 
 class ConteudoController extends ApiController
 {
@@ -143,9 +143,7 @@ class ConteudoController extends ApiController
         $conteudo->qt_downloads = $this->conteudo::INIT_COUNT;
         $conteudo->qt_access = $this->conteudo::INIT_COUNT;
 
-        if ($conteudo->save()) {
-            $this->createFile($conteudo->id, $this->request->download);
-        }
+        $conteudo->save();
         // PALAVRAS CHAVE, COMPONENTES CURRICULARES
         $conteudo->tags()->attach(explode(',', $this->request->tags));
         $conteudo->componentes()->attach(explode(',', $this->request->componentes));
@@ -158,13 +156,15 @@ class ConteudoController extends ApiController
     private function saveOptions($conteudo_id)
     {
         $tipo = collect(DB::select('select * from tipos where id = ?', [$this->request->tipo_id]))->first();
+        $arr_file = $this->createFile($conteudo->id, $this->request->download);
         $options = [
             'tipo'          => $tipo->id,
             'componentes'   => [$this->request->componentes],
             'tags'          => [$this->request->tags],
             'site'          => $this->request->site, // URL DO SITE
             'guia'          => $this->request->guia, // ARQUIVO DE GUIA PEDAGOGICA
-            'download'      => $this->request->hasFile('download') ? $conteudo_id . '.' . $this->request->download->guessExtension() : null,
+            //'download'      => $this->request->hasFile('download') ? $conteudo_id . '.' . $this->request->download->guessExtension() : null,
+            'download'      => $this->request->hasFile('download') ? $arr_file : null,
             'visualizacao'  => null, // ARQUIVO DE VISUALIZACAO - $this->request->visualizacao
         ];
         $conteudo = $this->conteudo::findOrFail($conteudo_id);
@@ -201,7 +201,7 @@ class ConteudoController extends ApiController
         $conteudo->ts_documento = $fullTextSearch->ts_documento;
         $conteudo->save();
     }
-
+    public function lerHDImage(){}
     /**
      * Atualiza o conteúdo.
      *
@@ -210,11 +210,16 @@ class ConteudoController extends ApiController
      */
     public function update($id)
     {
+        $conteudo = $this->conteudo::find($id);
+
+        if (Gate::denies('update', $conteudo)) {
+            return $this->errorResponse([], 'Usuário sem permissão de acesso!', 403);
+        }
+
         $validator = Validator::make($this->request->all(), config("rules.conteudo"));
         if ($validator->fails()) {
             return $this->errorResponse($validator->errors(), "Não foi possível atualizar o conteúdo", 201);
         }
-        $conteudo = $this->conteudo::find($id);
         $conteudo->fill($this->request->all());
         if ($conteudo->save()) {
             $this->createFile($id, $this->request->download);
@@ -238,6 +243,14 @@ class ConteudoController extends ApiController
     public function delete($id)
     {
         $conteudo = $this->conteudo::with('tags')->find($id);
+        // or Gate::denies('super-admin', $conteudo)
+        // if (Gate::denies('super-admin', $conteudo)) {
+        //     return $this->errorResponse([], 'Usuário sem permissão de acesso!', 403);
+        // }
+        if (Gate::denies('delete', $conteudo)) {
+            return $this->errorResponse([], 'Usuário sem permissão de acesso!', 403);
+        }
+
         $conteudo->tags()->detach();
         $conteudo->componentes()->detach();
         $conteudo->niveis()->detach();
